refactor(stella-cli): subsession's tests move to subsession/tests.rs
The SUB-AGENTS overlay (#4334) grew subsession.rs to 1545 lines, a
first-time crossing of the 1500-line ratchet that the baseline refuses
to absorb. The test module moves to a sibling file — driver/tests.rs is
the pattern — and the one code comment whose example path read to
check-doc-links as a citation of a missing document names a bare file
instead.

Refs #4334
